temporary commit test page
I am temporarily commiting a test post page to test anchor links. When
I create links in to headings or footnotes within the same document
and browse the post from localhost the links are being resolved
relative to localhost root rather than relative to the current
document. I want to find out if this is only an issue when browsing
locally. I will be deleting this post after I have finished testing.