Completed the series on encapsulating dependencies in a docker image
Finished the post titled Encapsulate Angular WebExtension Dependencies
In a Docker Image.

Updated the previous article with a prerequisite section.

Created a post directory under /static so that I can keep images in
static paths that are otherwise the same as the path to the published
post.

I renamed the first article in the series to Separate Dependencies
From Implementation Using POSIX ACL Extended Attributes & Docker, but
kept the path the same. I thought the new title more accurately
describes the subject being discussed.